requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
Pillow>=10.0.0,<12.0.0
ImageHash>=4.3.1,<5.0.0
python-multipart>=0.0.9,<0.1.0
httpx>=0.27.0,<1.0.0

